refactor (#830)
