+++ src/modules/tests/candidate/__tests__/candidate.reducer.spec.ts
@@ -1,14 +1,11 @@
-import { Candidate } from '@dvsa/mes-journal-schema';
 import { candidateReducer } from '../candidate.reducer';
 import { PopulateCandidateDetails } from '../candidate.actions';
+import { candidateMock } from '../../__mocks__/tests.mock';
 
 describe('candidate reducer', () => {
   it('should return the candidate from a start test action', () => {
-    const mockCandidate: Candidate = {
-      candidateId: 123,
-    };
-    const result = candidateReducer(null, new PopulateCandidateDetails(mockCandidate));
+    const result = candidateReducer(null, new PopulateCandidateDetails(candidateMock));
 
-    expect(result).toBe(mockCandidate);
+    expect(result).toBe(candidateMock);
   });
 });
