--- src/shared/mocks/candidate.mock.ts
@@ -1,9 +1,0 @@
-import { Candidate } from '@dvsa/mes-test-schema/categories/B';
-
-export const candidateMock: Candidate = {
-  driverNumber: '123',
-  candidateName: {
-    firstName: 'Joe',
-    lastName: 'Bloggs',
-  },
-};
